src/requirements.txt
dill==0.3.1.1
google-cloud-storage==1.26.0
google-cloud-bigquery==1.24.0
gspread==3.3.0
kedro[spark]==0.16.2
ipython>=7.0.0, <8.0
jupyter>=1.0.0, <2.0
jupyterlab==0.31.1
matplotlib==3.3.0
nbstripout==0.3.3
numpy>=1.9.3
networkx>=2.2
nbconvert==5.6.1
oauth2client
pandas==0.25.3
pandas-profiling==2.7.1
plotly
python-dotenv==0.11.0
pytest==5.3.5
pytest-cov>=2.5, <3.0
pytest-mock>=1.7.1,<2.0
seaborn==0.9.0
statsmodels>=0.11.0
scikit-learn==0.21.3
scipy==1.5.2
shap==0.30.1
wheel==0.32.2
xgboost==0.90
gcsfs==0.4.0
cx_oracle
patsy
pyspark==2.4.5
hdfs==2.5.8
pydotplus
argcomplete
parquet
pyarrow
